исправлены замечания кодревьера
