+++ gateway/src/main/java/ru/practicum/shareit/ShareItGateway.java
@@ -5,8 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class ShareItGateway {
-	public static void main(String[] args) {
-		SpringApplication.run(ShareItGateway.class, args);
-	}
-
+    public static void main(String[] args) {
+        SpringApplication.run(ShareItGateway.class, args);
+    }
 }
