+++ gateway/src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -8,10 +8,12 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.booking.dto.BookItemRequestDto;
 import ru.practicum.shareit.booking.dto.BookingState;
+import ru.practicum.shareit.exception.ValidationException;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
+import java.time.LocalDateTime;
 
 @Controller
 @RequestMapping(path = "/bookings")
@@ -22,10 +24,12 @@ public class BookingController {
     private final BookingClient bookingClient;
 
     @GetMapping
-    public ResponseEntity<Object> getBookings(@RequestHeader("X-Sharer-User-Id") long userId,
-                                              @RequestParam(name = "state", defaultValue = "all") String stateParam,
-                                              @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
-                                              @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
+    public ResponseEntity<Object> getBookings(
+            @RequestHeader("X-Sharer-User-Id") long userId,
+            @RequestParam(name = "state", defaultValue = "all") String stateParam,
+            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
+            @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
+        validatePage(from, size);
         BookingState state = BookingState.from(stateParam)
                 .orElseThrow(() -> new IllegalArgumentException("Unknown state: " + stateParam));
         log.info("Get booking with state {}, userId={}, from={}, size={}", stateParam, userId, from, size);
@@ -36,6 +40,7 @@ public class BookingController {
     public ResponseEntity<Object> bookItem(@RequestHeader("X-Sharer-User-Id") long userId,
                                            @RequestBody @Valid BookItemRequestDto requestDto) {
         log.info("Creating booking {}, userId={}", requestDto, userId);
+        validateBooking(requestDto);
         return bookingClient.bookItem(userId, requestDto);
     }
 
@@ -54,12 +59,39 @@ public class BookingController {
     }
 
     @GetMapping("/owner")
-    public ResponseEntity<Object> findOwnerBookings(@RequestHeader("X-Sharer-User-Id") long userId,
-                                                    @RequestParam(name = "state", defaultValue = "all") String stateParam,
-                                                    @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
-                                                    @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
+    public ResponseEntity<Object> findOwnerBookings(
+            @RequestHeader("X-Sharer-User-Id") long userId,
+            @RequestParam(name = "state", defaultValue = "all") String stateParam,
+            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
+            @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
+        validatePage(from, size);
         BookingState state = BookingState.from(stateParam)
                 .orElseThrow(() -> new IllegalArgumentException("Unknown state: " + stateParam));
         return bookingClient.findOwnerBookings(userId, state, from, size);
+    }
+
+    private void validateBooking(BookItemRequestDto bookingDto) {
+        if (bookingDto.getStart().isBefore(LocalDateTime.now())) {
+            throw new ValidationException(String.format(
+                    "Дата начала брони (%s) находится в прошлом", bookingDto.getStart()));
+        } else if (bookingDto.getEnd().isBefore(LocalDateTime.now())) {
+            throw new ValidationException(String.format(
+                    "Дата окончания брони (%s) находится в прошлом", bookingDto.getEnd()));
+        } else if (bookingDto.getEnd().isBefore(bookingDto.getStart())) {
+            throw new ValidationException(String.format(
+                    "Дата окончания брони (%s) раньше даты начала (%s)",
+                    bookingDto.getEnd(), bookingDto.getStart()));
+        }
+    }
+
+    private int validatePage(int from, int size) {
+        if (size <= 0) {
+            throw new ValidationException(String.format("Параметр size (%s) задан некорректно", size));
+        }
+        if (from < 0) {
+            throw new ValidationException(String.format("Параметр from (%s) задан некорректно", from));
+        }
+        int page = from / size;
+        return page;
     }
 }
