+++ gateway/src/main/java/ru/practicum/shareit/booking/dto/BookingState.java
@@ -5,25 +5,25 @@ import ru.practicum.shareit.exception.ValidationException;
 import java.util.Optional;
 
 public enum BookingState {
-	// Все
-	ALL,
-	// Текущие
-	CURRENT,
-	// Будущие
-	FUTURE,
-	// Завершенные
-	PAST,
-	// Отклоненные
-	REJECTED,
-	// Ожидающие подтверждения
-	WAITING;
+    // Все
+    ALL,
+    // Текущие
+    CURRENT,
+    // Будущие
+    FUTURE,
+    // Завершенные
+    PAST,
+    // Отклоненные
+    REJECTED,
+    // Ожидающие подтверждения
+    WAITING;
 
-	public static Optional<BookingState> from(String stringState) {
-		for (BookingState state : values()) {
-			if (state.name().equalsIgnoreCase(stringState)) {
-				return Optional.of(state);
-			}
-		}
-		throw new ValidationException(String.format("Unknown state: %s", stringState));
-	}
+    public static Optional<BookingState> from(String stringState) {
+        for (BookingState state : values()) {
+            if (state.name().equalsIgnoreCase(stringState)) {
+                return Optional.of(state);
+            }
+        }
+        throw new ValidationException(String.format("Unknown state: %s", stringState));
+    }
 }
