+++ gateway/src/main/resources/application.properties
@@ -1,7 +1,2 @@
 logging.level.org.springframework.web.client.RestTemplate=DEBUG
-#logging.level.org.apache.http=DEBUG
-#logging.level.httpclient.wire=DEBUG
-
-server.port=8080
-
 shareit-server.url=http://localhost:9090
