+++ server/src/main/java/ru/practicum/shareit/ShareItServer.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class ShareItServer {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ShareItServer.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ShareItServer.class, args);
+    }
 
 }
