+++ server/src/main/resources/application.properties
@@ -1,20 +1,16 @@
 server.port=9090
-
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
 spring.jpa.properties.hibernate.format_sql=true
 spring.sql.init.mode=always
-
 #---
 spring.datasource.driverClassName=org.postgresql.Driver
 spring.datasource.url=jdbc:postgresql://localhost:6541/shareit
 spring.datasource.username=root
 spring.datasource.password=root
-
-# настройка автоматического создания таблиц при помощи sql-скрипта
+# РЅР°СЃС‚СЂРѕР№РєР° Р°РІС‚РѕРјР°С‚РёС‡РµСЃРєРѕРіРѕ СЃРѕР·РґР°РЅРёСЏ С‚Р°Р±Р»РёС† РїСЂРё РїРѕРјРѕС‰Рё sql-СЃРєСЂРёРїС‚Р°
 javax.persistence.schema-generation.database.action=create
 javax.persistence.schema-generation.create-script-source=schema.sql
-
 #---
 spring.config.activate.on-profile=ci,test
 spring.datasource.driverClassName=org.h2.Driver
